# docs/swagger.yaml
basePath: /api
definitions:
  req.RatePackageItem:
    properties:
      date:
        type: string
      price:
        type: number
      ratePlan:
        type: string
    type: object
...
  req.RequestCreateRoomType:
    properties:
      activated:
        type: boolean
      air_conditional:
        type: boolean
      bathroom_nums:
        type: integer
      bay:
        type: boolean
      bbq:
        type: boolean
      bed_nums:
        type: integer
      city:
        type: boolean
      description:
        type: string
      desk:
        type: boolean
      fruit:
        type: boolean
      fryer:
        type: boolean
      garden:
        type: boolean
      hairdry:
        type: boolean
      heater:
        type: boolean
      hotel_id:
        type: string
      id:
        type: string
      iron:
        type: boolean
      kitchen:
        type: boolean
      kitchen_tool:
        type: boolean
      lake:
        type: boolean
      max_adult:
        type: integer
      max_children:
        type: integer
      mountain:
        type: boolean
      name:
        type: string
      ocean:
        type: boolean
      photos:
        type: string
      private_balcony:
        type: boolean
      private_pool:
        type: boolean
      river:
        type: boolean
      shower:
        type: boolean
      slippers:
        type: boolean
      sofa:
        type: boolean
      soundproof:
        type: boolean
      tivi:
        type: boolean
      toiletries:
        type: boolean
      towels:
        type: boolean
      wine:
        type: boolean
    type: object
...
  req.RequestUpdateRatePackage:
    properties:
      data:
        items:
          $ref: '#/definitions/req.RatePackageItem'
        type: array
    type: object
...
  req.RequestUpdateRoomNight:
    properties:
      quantity:
        type: integer
      room_type_id:
        type: string
      selected_date:
        type: string
    type: object
  req.RequestUpdateRoomType:
    properties:
      activated:
        type: boolean
      air_conditional:
        type: boolean
      bathroom_nums:
        type: integer
      bay:
        type: boolean
      bbq:
        type: boolean
      bed_nums:
        type: integer
      city:
        type: boolean
      description:
        type: string
      desk:
        type: boolean
      fryer:
        type: boolean
      fuirt:
        type: boolean
      garden:
        type: boolean
      hairdry:
        type: boolean
      heater:
        type: boolean
      iron:
        type: boolean
      is_delete:
        type: boolean
      kitchen:
        type: boolean
      kitchen_tool:
        type: boolean
      lake:
        type: boolean
      max_adult:
        type: integer
      max_children:
        type: integer
      mountain:
        type: boolean
      name:
        type: string
      ocean:
        type: boolean
      private_balcony:
        type: boolean
      private_pool:
        type: boolean
      river:
        type: boolean
      sea:
        type: boolean
      shower:
        type: boolean
      slippers:
        type: boolean
      sofa:
        type: boolean
      soundproof:
        type: boolean
      tivi:
        type: boolean
      toiletries:
        type: boolean
      towels:
        type: boolean
      wine:
        type: boolean
    type: object
  res.Response:
    properties:
      data: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
host: https://siphoria-api-production.up.railway.app
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server Petstore server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: HOTEL BOOKING API
  version: "1.0"
paths:
...
  /rooms:
    post:
      consumes:
      - application/json
      parameters:
      - description: room
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/req.RequestCreateRoomType'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/res.Response'
      summary: Save room type
      tags:
      - room-service
  /rooms/:id:
    patch:
      consumes:
      - application/json
      parameters:
      - description: RoomType
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/req.RequestUpdateRoomType'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/res.Response'
      summary: update room type
      tags:
      - room-service
  /rooms/:id/photos:
    patch:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/res.Response'
      summary: Update room photos
      tags:
      - room-service
  /rooms/ratepackages:
    post:
      consumes:
      - application/json
      parameters:
      - description: room
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/req.RequestUpdateRatePackage'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/res.Response'
      summary: Update ratepackages
      tags:
      - room-service
  /rooms/roomnights:
    post:
      consumes:
      - application/json
      parameters:
      - description: room
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/req.RequestUpdateRoomNight'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/res.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/res.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/res.Response'
      summary: Update room nights
      tags:
      - room-service
...
swagger: "2.0"